profile get,post endpoints created

+++ src/main/java/com/sandeep/portfolio/repository/Contactrepo.java
@@ -3,5 +3,5 @@ package com.sandeep.portfolio.repository;
 import com.sandeep.portfolio.entity.Contact;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Contactrepo extends JpaRepository<Contact,Integer> {
+public interface Contactrepo extends JpaRepository<Contact,Long> {
 }
