+++ src/main/java/com/sandeep/portfolio/repository/Mediarepo.java
@@ -3,5 +3,5 @@ package com.sandeep.portfolio.repository;
 import com.sandeep.portfolio.entity.Media;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Mediarepo extends JpaRepository<Media,Integer> {
+public interface Mediarepo extends JpaRepository<Media,Long> {
 }
