+++ src/main/java/com/sandeep/portfolio/repository/Profilerepo.java
@@ -3,5 +3,5 @@ package com.sandeep.portfolio.repository;
 import com.sandeep.portfolio.entity.Profile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Profilerepo extends JpaRepository<Profile,Integer> {
+public interface Profilerepo extends JpaRepository<Profile,Long> {
 }
