+++ src/main/java/com/sandeep/portfolio/repository/Projectrepo.java
@@ -3,5 +3,5 @@ package com.sandeep.portfolio.repository;
 import com.sandeep.portfolio.entity.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Projectrepo extends JpaRepository<Project,Integer> {
+public interface Projectrepo extends JpaRepository<Project,Long> {
 }
